+++ processor/src/stack/mod.rs
@@ -98,7 +98,7 @@ impl Stack {
 
     /// Returns a copy of the item currently at the top of the stack.
     pub fn peek(&self) -> Felt {
-        self.trace.peek_at(self.clk)
+        self.trace.peek_at(self.clk as u32)
     }
 
     /// Return n items from the top of the stack, including the values from the overflow table.
@@ -109,7 +109,9 @@ impl Stack {
         let num_items = cmp::min(n, self.depth());
 
         let num_top_items = cmp::min(MIN_STACK_DEPTH, num_items);
-        let mut result = self.trace.get_stack_values_at(self.clk, num_top_items);
+        let mut result = self
+            .trace
+            .get_stack_values_at(self.clk as u32, num_top_items);
 
         if num_items > MIN_STACK_DEPTH {
             let num_overflow_items = num_items - MIN_STACK_DEPTH;
@@ -155,7 +157,7 @@ impl Stack {
     /// Returns stack state at the specified clock cycle.
     ///
     /// This includes the stack + overflow entries.
-    pub fn get_state_at(&self, clk: usize) -> Vec<Felt> {
+    pub fn get_state_at(&self, clk: u32) -> Vec<Felt> {
         let mut result = self.trace.get_stack_state_at(clk).to_vec();
         self.overflow.append_state_into(&mut result, clk);
 
@@ -168,13 +170,14 @@ impl Stack {
     /// Returns the value located at the specified position on the stack at the current clock cycle.
     pub fn get(&self, pos: usize) -> Felt {
         debug_assert!(pos < MIN_STACK_DEPTH, "stack underflow");
-        self.trace.get_stack_value_at(self.clk, pos)
+        self.trace.get_stack_value_at(self.clk as u32, pos)
     }
 
     /// Sets the value at the specified position on the stack at the next clock cycle.
     pub fn set(&mut self, pos: usize, value: Felt) {
         debug_assert!(pos < MIN_STACK_DEPTH, "stack underflow");
-        self.trace.set_stack_value_at(self.clk + 1, pos, value);
+        self.trace
+            .set_stack_value_at(self.clk as u32 + 1, pos, value);
     }
 
     /// Copies stack values starting at the specified position at the current clock cycle to the
@@ -184,7 +187,7 @@ impl Stack {
             start_pos < MIN_STACK_DEPTH,
             "start cannot exceed stack top size"
         );
-        self.trace.copy_stack_state_at(self.clk, start_pos);
+        self.trace.copy_stack_state_at(self.clk as u32, start_pos);
     }
 
     /// Copies stack values starting at the specified position at the current clock cycle to
@@ -204,17 +207,18 @@ impl Stack {
             0..=MAX_TOP_IDX => unreachable!("stack underflow"),
             MIN_STACK_DEPTH => {
                 // Shift in a ZERO, to prevent depth shrinking below the minimum stack depth.
-                self.trace.stack_shift_left_at(self.clk, start_pos, ZERO);
-                self.trace.copy_helpers_at(self.clk);
+                self.trace
+                    .stack_shift_left_at(self.clk as u32, start_pos, ZERO);
+                self.trace.copy_helpers_at(self.clk as u32);
             }
             _ => {
                 // Update the stack & overflow table.
-                let (from_overflow, prev_addr) = self.overflow.pop(self.clk);
+                let (from_overflow, prev_addr) = self.overflow.pop(self.clk as u32);
                 self.trace
-                    .stack_shift_left_at(self.clk, start_pos, from_overflow);
+                    .stack_shift_left_at(self.clk as u32, start_pos, from_overflow);
 
                 // Update the bookkeeping & helper columns.
-                self.trace.helpers_shift_left_at(self.clk, prev_addr);
+                self.trace.helpers_shift_left_at(self.clk as u32, prev_addr);
 
                 // Stack depth only decreases when it is greater than the minimum stack depth.
                 self.depth -= 1;
@@ -233,14 +237,14 @@ impl Stack {
         );
 
         // Update the stack.
-        self.trace.stack_shift_right_at(self.clk, start_pos);
+        self.trace.stack_shift_right_at(self.clk as u32, start_pos);
 
         // Update the overflow table.
-        let to_overflow = self.trace.get_stack_value_at(self.clk, MAX_TOP_IDX);
-        self.overflow.push(to_overflow, self.clk);
+        let to_overflow = self.trace.get_stack_value_at(self.clk as u32, MAX_TOP_IDX);
+        self.overflow.push(to_overflow, self.clk as u32);
 
         // Update the bookkeeping & helper columns.
-        self.trace.helpers_shift_right_at(self.clk);
+        self.trace.helpers_shift_right_at(self.clk as u32);
 
         // Stack depth always increases on right shift.
         self.depth += 1;
@@ -258,7 +262,7 @@ impl Stack {
     ///
     /// Trace length is doubled every time it needs to be increased.
     pub fn ensure_trace_capacity(&mut self) {
-        self.trace.ensure_trace_capacity(self.clk);
+        self.trace.ensure_trace_capacity(self.clk as u32);
     }
 
     // TEST HELPERS
@@ -270,6 +274,6 @@ impl Stack {
     /// the stack depth is less than 16, the un-used slots contain ZEROs.
     #[cfg(test)]
     pub fn trace_state(&self) -> StackTopState {
-        self.trace.get_stack_state_at(self.clk)
+        self.trace.get_stack_state_at(self.clk as u32)
     }
 }
